+++ plugins/gui/src/keybind_edit/keybind_edit.cpp
@@ -1,22 +1,38 @@
 #include "gui/keybind_edit/keybind_edit.h"
+#include "gui/settings/settings_items/settings_item_keybind.h"
 
 #include <QEvent>
 #include <QKeyEvent>
-#include <QMessageBox>
+#include <QStyle>
 
 namespace hal
 {
-    KeybindEdit::KeybindEdit(QWidget* parent): QKeySequenceEdit(parent)
+    KeybindEdit::KeybindEdit(QWidget* parent)
+        : QKeySequenceEdit(parent), mItem(nullptr),
+          mSkipValidate(false), mValidated(true), mGrab(false)
     {;}
 
-    void KeybindEdit::addValidator(Validator* v)
+    void KeybindEdit::setHasGrab(bool isgrab)
     {
-        mValidator.addValidator(v);
+        if (isgrab==mGrab) return;
+        mGrab = isgrab;
+
+        setStyleSheet(QString("background: rgb(0,0,%1);").arg(mGrab ? 40 : 0));
+
+        QStyle* s = style();
+        s->unpolish(this);
+        s->polish(this);
+
     }
 
-    void KeybindEdit::setValidated(bool validated)
+    void KeybindEdit::setValidated(bool valid)
     {
-       mValidated = validated;
+       mValidated = valid;
+    }
+
+    bool KeybindEdit::hasGrab()
+    {
+        return mGrab;
     }
 
     bool KeybindEdit::validated()
@@ -24,56 +40,77 @@ namespace hal
         return mValidated;
     }
 
-    void KeybindEdit::revalidate()
+    bool KeybindEdit::doValidate()
     {
-        setValidated(mValidator.validate(keySequence().toString()));
+        if (mSkipValidate) return true;
+        QKeySequence current = keySequence();
+        SettingsItemKeybind* item = AssignedKeybindMap::instance()->currentAssignment(current);
+        bool ok = (item == nullptr || item == mItem);
+
+        setValidated(ok);
+        if (ok)
+        {
+            AssignedKeybindMap::instance()->tempAssign(current,mItem,mOldSequence);
+            mOldSequence = current;
+            Q_EMIT(editAccepted());
+        }
+        else
+        {
+            // revert
+            Q_EMIT(editRejected(
+                        QString("<%1> is already assigned:\n<%2>")
+                        .arg(current.toString())
+                        .arg(item->label())));
+            restoreOldSequence();
+        }
+        return ok;
+    }
+
+    void KeybindEdit::load(const QKeySequence& seq, SettingsItemKeybind* item)
+    {
+        mOldSequence = seq;
+        mItem = item;
+        QKeySequenceEdit::setKeySequence(seq);
+    }
+
+    void KeybindEdit::restoreOldSequence()
+    {
+        mSkipValidate = true;
+        setKeySequence(mOldSequence);
+        mSkipValidate = false;
     }
 
     bool KeybindEdit::event(QEvent* e)
     {
-        bool recognized = false;
+        bool recognized = true;
         switch(e->type())
         {
-            /*
+        /*
              * Make sure we have the keyboard exclusively.
              * This suppresses defined keyboard shortcuts
              * and lets us react properly if the user
              * attempts to configure stupid keybinds like
              * Alt-F4.
              */
-            case QEvent::FocusIn:
-                grabKeyboard();
-                recognized = true;
-                break;
-            case QEvent::FocusOut:
-                // FIXME this messes with the API. Better define a Validator that
-                // can actually handle QVariants and thus QKeySequences.
-                if (!mValidator.validate(keySequence().toString()))
-                {
-                    // revert
-                    QKeySequence failed = keySequence();
-                    setKeySequence(mOldSequence);
-                    if (failed != mFailedValidate)
-                    {
-                        QMessageBox::warning(this, "Rejected!", mValidator.failText());
-                        mFailedValidate = failed;
-                    }
-                    Q_EMIT(editRejected());
-                }
-                else
-                {
-                    mOldSequence = keySequence();
-                }
-                releaseKeyboard();
-                recognized = true;
-                break;
-            case QEvent::KeyRelease:
-                // FIXME this messes with the API. Better define a Validator that
-                // can actually handle QVariants and thus QKeySequences.
-                revalidate();
-                recognized = true;
-                break;
-            default: break; // suppress -Wswitch compiler warning
+        case QEvent::FocusIn:
+            grabKeyboard();
+            setHasGrab(true);
+            break;
+        case QEvent::FocusOut:
+            // FIXME this messes with the API. Better define a Validator that
+            // can actually handle QVariants and thus QKeySequences.
+            doValidate();
+            setHasGrab(false);
+            releaseKeyboard();
+            break;
+        case QEvent::KeyRelease:
+            // FIXME this messes with the API. Better define a Validator that
+            // can actually handle QVariants and thus QKeySequences.
+            doValidate();
+            break;
+        default:
+            recognized = false;
+            break;
         }
         recognized |= QKeySequenceEdit::event(e);
         return recognized;