+++ plugins/gui/src/settings/settings_items/settings_item_keybind.cpp
@@ -5,7 +5,25 @@
 
 namespace hal
 {
-    QMap<QKeySequence,SettingsItemKeybind*>* SettingsItemKeybind::sKeysAssigned;
+    AssignedKeybindMap* AssignedKeybindMap::inst = nullptr;
+
+    AssignedKeybindMap* AssignedKeybindMap::instance()
+    {
+        if (!inst) inst = new AssignedKeybindMap;
+        return inst;
+    }
+
+    void AssignedKeybindMap::registerKeybind(const QKeySequence &key, SettingsItemKeybind *setting)
+    {
+        mAssignedMap.insert(key,setting);
+    }
+
+    void AssignedKeybindMap::tempAssign(const QKeySequence& newkey, SettingsItemKeybind *setting, const QKeySequence& oldkey)
+    {
+        if (!oldkey.isEmpty())
+            mTempMap.remove(oldkey);
+        mTempMap.insert(newkey,setting);
+    }
 
     SettingsItemKeybind::SettingsItemKeybind(const QString& label, const QString& tag, const QKeySequence& defVal, const QString& cat, const QString& descr, bool isGlobal)
     {
@@ -18,7 +36,7 @@ namespace hal
         mIsGlobal = isGlobal;
         
         SettingsManager::instance()->registerSetting(this);
-        registerKeybind(mValue,this);
+        AssignedKeybindMap::instance()->registerKeybind(mValue,this);
     }
 
     QVariant SettingsItemKeybind::value() const
@@ -29,12 +47,6 @@ namespace hal
     QVariant SettingsItemKeybind::defaultValue() const
     {
         return mDefaultValue;
-    }
-
-    void SettingsItemKeybind::registerKeybind(const QKeySequence &key, SettingsItemKeybind *setting)
-    {
-        if (!sKeysAssigned) sKeysAssigned = new QMap<QKeySequence,SettingsItemKeybind*>();
-        sKeysAssigned->insert(key,setting);
     }
 
     void SettingsItemKeybind::setDefaultValue(const QVariant& dv)
@@ -57,13 +69,7 @@ namespace hal
         if(mValue == newValue)
             return;
 
-        if (sKeysAssigned)
-        {
-            auto it = sKeysAssigned->find(mValue);
-            if (it != sKeysAssigned->end()) sKeysAssigned->erase(it);
-        }
         mValue = newValue;
-        registerKeybind(mValue,this);
 
         Q_EMIT valueChanged();
         Q_EMIT keySequenceChanged(v.toString());
@@ -72,11 +78,5 @@ namespace hal
     SettingsWidget* SettingsItemKeybind::editWidget(QWidget* parent)
     {
         return new SettingsWidgetKeybind(this, parent);
-    }
-
-    SettingsItemKeybind *SettingsItemKeybind::currentKeybindAssignment(const QKeySequence &needle)
-    {
-        if (!sKeysAssigned) return nullptr;
-        return sKeysAssigned->value(needle);
     }
 }